Moved InstantWrapperColumnType to Java8TimeColumns, direct use of BooleanWrapper
