+++ runtime/src/main/kotlin/pl/touk/krush/WrapperColumn.kt
@@ -1,8 +1,6 @@
 package pl.touk.krush
 
 import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.`java-time`.JavaInstantColumnType
-import org.jetbrains.exposed.sql.vendors.currentDialect
 import java.math.BigDecimal
 import java.time.Instant
 import kotlin.reflect.KClass
@@ -78,20 +76,6 @@ class StringWrapperColumnType<out Wrapper : Any>(
     }
 }
 
-class InstantWrapperColumnType<out Wrapper : Any>(
-        wrapperClazz: KClass<Wrapper>,
-        instanceCreator: (Instant) -> Wrapper,
-        valueExtractor: (Wrapper) -> Instant
-) : WrapperColumnType<Instant, Wrapper>(JavaInstantColumnType(), Instant::class, wrapperClazz, instanceCreator, valueExtractor) {
-
-    override fun valueFromDB(value: Any) = when (value) {
-        // Supporting same types as org.jetbrains.exposed.sql.`java-time`.JavaInstantColumnType
-        is java.sql.Timestamp -> instanceCreator(rawColumnType.valueFromDB(value) as Instant)
-        is String -> instanceCreator(rawColumnType.valueFromDB(value) as Instant)
-        else -> error("Database value $value of class ${value::class.qualifiedName} is not valid $rawClazz")
-    }
-}
-
 class BooleanWrapperColumnType<out Wrapper : Any>(
         wrapperClazz: KClass<Wrapper>,
         instanceCreator: (Boolean) -> Wrapper,
@@ -115,14 +99,6 @@ inline fun <reified Wrapper : Any> Table.stringWrapper(
         noinline valueExtractor: (Wrapper) -> String
 ): Column<Wrapper> = registerColumn(
         name, StringWrapperColumnType(Wrapper::class, instanceCreator, valueExtractor)
-)
-
-inline fun <reified Wrapper : Any> Table.instantWrapper(
-        name: String,
-        noinline instanceCreator: (Instant) -> Wrapper,
-        noinline valueExtractor: (Wrapper) -> Instant
-): Column<Wrapper> = registerColumn(
-        name, InstantWrapperColumnType(Wrapper::class, instanceCreator, valueExtractor)
 )
 
 inline fun <reified Wrapper : Any> Table.booleanWrapper(
